+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "repoTS" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1] - 2025-04-17
 
-- Initial release
+### Added
+- Initial release of repoTS
+- Automatic TypeScript SDK detection based on the current file being edited
+- Workspace setting updates for typescript.tsdk
+- Reload prompts when SDK changes